Scalable Semantic Search System using Qdrant + FastAPI (HNSW Indexing)/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.43.0",
]
